Create Lab 6
I am aware that I did not get problem 4 correct however, I spent almost 4 hours trying and failing to correct the issue, and I cannot see a solution.  Either I am too dense to see the obvious, or I am not looking at the right material.  Can you give me a direction as to when to look to fix the gap in my knowledge base?